requirements.txt
playwright==1.40.0
pandas==2.1.4
asyncpg==0.29.0
python-dotenv==1.0.0
openpyxl==3.1.2
